Added documentation and fixed/added some more methods
Firstly, I added documentation to some functions and variables so that it explains what they do and what they return. I just like to do it. I then fixed and changed up some functions so that they would work better such as ConstructTiles and ConstructGrid. The few new methods i added were just for like changing private variables such as numTilesTaken. I just added two referenes for the player and grid in the GameBehaviorScript and the gameobject itself. Not too much either, I am gonna wait a bit for the grid tile sprites to then use to construct the grid for testing before i go for more of the functionality.