added more scripts. Grid/board constructs and so do player tiles
I added two prefabs: one for the BoardSquare and another for a GamePiece. I added scripts to both of them with not code in them yet. Except BoardSquare which has a method to change the sprite of the block, This is just used for the board creation right now (to make alternating pattern). I finalized the constructPieces for the player, so now depending on the size, the right number of pieces will be created for the player and added as children to the player object. I also finalized the createGrid() funciton in the Grid class that will fully construct and internal array for the grid/game state, and there is another array created for the visual board. They are both initialized in the same function.I added some extra get functions and then tested it out in the Game class. It seemed to work when I pressed play so I will call that a success for right now. Next is to make the player pieces go to the right squares, then after that is to be able to add behavior to the pieces. Also added some documentation and debug logs to make sure things work out.