Added to piecebhavior script and displayed pieces from player
I added most of the stuff from the jamboard onto the PieceBehaviorScript so that the pieces would have the correct index and other stuff. The behavior for moving the piece and stuff is not implemented yet, gonna wait a bit on that to work on with Tyler, but otherwise it is working. I editted the GridBehavior and GameManagerScript to display the pieces properly. Firstly, the pieces are grabbed once in the ConstructGrid method and then properly assigned to the grid this time, it was not working correctly before. I did this by just implementing a simple counter rather than calculating it every time. Then in the game manager script I just checked every boardsquare from the grid(not visual grid) and then if it was not null, then i moved that specific tile to the position which it corresponds to. This was just done by setting its position equal to the position of the specific tile at position (i, j). There were some changes to the loop in Grid Behavior, more specifically with the order of i and j, but I might mess with this a bit to get it how i want it. Lastly, I just added some documentation to the new PieceBehaviorScript and to the GridBehvaiorScript and capitalized some method names. I think I am wanting to move the displaying of the pieces from the gamemanager script to the constructgrid method because it makes more sense there. I might have some trouble because the if statements and stuff but i think it should generally be fine to implement. Next is to align the pieces to the correct side or maybe work on where the origin of the grid will be located, then work on creating the methods to change the grid and to check which squares are available to that specific piece.