Fixed coordinate system, started work on grid methods
I took some time to rework the grid a bit. Currently the board tiles were rendered so that (0,0) was at the top right. This was better not necessarily ideal because the game pieces spawned from right to left starting at the top right. So I reworked the for loops, i just made it so it looped through j first instead of i. It didn't change anything until Ichanged the position of where the board tiles would spawn and then everything reoriented properly. honestly was just a simple fix. I started to work on the MovePiece method in the grid class, but then realized I also needed the available pieces method (or maybe not depending on implementation of piece/player logic). So I just made a method that would change the position of the current piece on the grid to the one inputted in the method. I also have to change the position variable in Piece to be Vector2Int since I decided the position would be relative to the board. Up next is just completing the method along with the GetAvailableTiles method and also creting a method to keep the piece in the new place it was just dropped.