Added GetAvailablePieces and tweaked MovePiece
I added the behavior for the getavailable pieces. It just checks if the place the piece was dropped is in bounds. if not then it returns false. It does not check if there are other pieces in the spot but I will add that later. I also added a few lines to movepiece() to use GetAvailablePieces behavior. I also added a position variable to the BoardSquareScript. This is because whenever we are using the location for pieces or the board it is a coordinate system realtive to the board. In the scene the location is different and only ruled by the boardsquares themselves. (so the actual transform is not messed with in code)