Added the ability to grab and move game objects
Added a script that gives the pieces collision, lets the player move them around with a mouse, and set them down. Makes a ray that checks if the player mouse is on top of a piece, and if the player clicks down, the piece snaps to the cursor position. The pieces all have collision so they do not overlap