Recreated Visual Script in C#
I rewrote the visual script in the Player class script to fine tune it a bit. I also realized that it is better to have the logic inside the Player script rather than the piecebehaviorscript. It looks neat for right now, but after some testing there is gonna be new logic to circumvent issues like overlapping pieces and the mouse exiting the collider of the piece (which makes the piece stop moving to the mouse position). Plan to fix this is making it so only the selected piece moves to the current mouse position as long as left mouse button held down (piece will probably stay selected unless mouse clicks elsewhere.