Changed Selection Logic
Changed so that the player script would just select the piece rather than move the piece. The player can also deselect by clicking out of a piece. The piece now moves through the piece script only if the mouse is held down and piece selected. The game script gives the proper position to each piece and now the piece will move to its proper position on board if let go during drag. Also the piece gets big when selected just for now. Next is to fix collision (so no collision between pieces) and to allow pieces to move to other squares (maybe highlighting).